* Large scale partial updates

* Large scale partial updates

* Fix merge test

* BLD: add partial/merge test to travis

* Accidentally deleted the header line on write_mat, and fixing travis config

* Curly braces...

* Seriously, missed a quote

* The prior if statement worked, travis was green, yet red

* Because it wasn't awk-ward already

* Addressing @antgonza's comment

+++ sucpp/unifrac.cpp
@@ -157,12 +157,12 @@ void su::stripes_to_condensed_form(std::vector<double*> &stripes, uint32_t n, do
     // n must be >= 2, but that should be enforced upstream as that would imply
     // computing unifrac on a single sample. 
 
-    uint32_t comb_N = comb_2(n);
+    uint64_t comb_N = comb_2(n);
     for(unsigned int stripe = start; stripe < stop; stripe++) {
         // compute the (i, j) position of each element in each stripe
-        uint32_t i = 0;
-        uint32_t j = stripe + 1;
-        for(uint32_t k = 0; k < n; k++, i++, j++) {
+        uint64_t i = 0;
+        uint64_t j = stripe + 1;
+        for(uint64_t k = 0; k < n; k++, i++, j++) {
             if(j == n) {
                 i = 0;
                 j = n - (stripe + 1);
@@ -170,7 +170,7 @@ void su::stripes_to_condensed_form(std::vector<double*> &stripes, uint32_t n, do
             // determine the position in the condensed form vector for a given (i, j)
             // based off of
             // https://docs.scipy.org/doc/scipy/reference/generated/scipy.spatial.distance.squareform.html
-            uint32_t comb_N_minus_i = comb_2(n - i);
+            uint64_t comb_N_minus_i = comb_2(n - i);
             cf[comb_N - comb_N_minus_i + (j - i - 1)] = stripes[stripe][k];
         }
     }
